[REFACTOR] #34 - Make calendarComponents
