+++ presentation/src/main/java/com/woowahan/accountbook/ui/calendar/CalendarScreen.kt
@@ -80,99 +80,16 @@ fun CalendarScreen(
                         CalendarItem(it)
                     }
                 }
+
                 Spacer(modifier = Modifier.height(20.dp))
-                Row(modifier = Modifier.padding(16.dp, 7.dp, 16.dp, 7.dp)) {
-                    Text(text = "수입", color = Purple, fontSize = 16.sp)
-                    Spacer(modifier = Modifier.weight(1f))
-                    Text(
-                        text = String.format("%,d", mainViewModel.totalIncome),
-                        color = Green6,
-                        fontSize = 16.sp
-                    )
-                }
+                CalendarText(text = "수입", amount = mainViewModel.totalIncome, color = Green6)
                 LightDivider(padding = 16)
-                Row(modifier = Modifier.padding(16.dp, 7.dp, 16.dp, 7.dp)) {
-                    Text(text = "지출", color = Red, fontSize = 16.sp)
-                    Spacer(modifier = Modifier.weight(1f))
-                    Text(
-                        text = String.format("%,d", mainViewModel.totalSpending),
-                        color = Red,
-                        fontSize = 16.sp
-                    )
-                }
+                CalendarText(text = "지출", amount = mainViewModel.totalSpending, color = Red)
                 LightDivider(padding = 16)
-                Row(modifier = Modifier.padding(16.dp, 7.dp, 16.dp, 7.dp)) {
-                    Text(text = "총합", color = Purple, fontSize = 16.sp)
-                    Spacer(modifier = Modifier.weight(1f))
-                    Text(
-                        text = String.format("%,d", mainViewModel.totalAmount),
-                        color = Purple,
-                        fontSize = 16.sp
-                    )
-                }
+                CalendarText(text = "총합", amount = mainViewModel.totalAmount, color = Purple)
                 Spacer(modifier = Modifier.height(10.dp))
                 BoldDivider()
             }
-        }
-    }
-}
-
-@Composable
-fun CalendarItem(calendar: Calendar) {
-    val modifier = Modifier
-        .fillMaxWidth()
-        .height(60.dp)
-        .background(if (calendar.isToday) White else Color.Transparent)
-        .drawBehind {
-            drawLine(
-                LightPurple,
-                Offset(0f, size.height),
-                Offset(size.width, size.height),
-                1.dp.value
-            )
-            drawLine(
-                LightPurple,
-                Offset(size.width, 0f),
-                Offset(size.width, size.height),
-                1.dp.value
-            )
-        }
-
-    Box(
-        modifier = modifier
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(4.dp)
-        ) {
-            Column {
-                if (calendar.income != "0") {
-                    Text(text = calendar.income, fontSize = 8.sp, color = Green6)
-                }
-                if (calendar.spending != "0") {
-                    Text(text = calendar.spending, fontSize = 8.sp, color = Red)
-                }
-                if (
-                    !(calendar.income == "0" &&
-                            calendar.spending == "0" &&
-                            calendar.total == "0") &&
-                    calendar.isCurrentMonth
-                )
-                    Text(text = calendar.total, fontSize = 8.sp, color = Purple)
-            }
-
-            Text(
-                modifier = Modifier.align(Alignment.BottomEnd),
-                text = calendar.day.toString(),
-                color = if (calendar.isCurrentMonth) {
-                    Purple
-                } else {
-                    Purple40
-                },
-                fontSize = 8.sp,
-                fontWeight = FontWeight.Bold
-            )
         }
     }
 }
